working cast with local audio script
vibe-coded the ever living hell out of this, use at own risk
it's easiest to just use catt for handling video and consider audio to
be a listener process that tries to keep up with video. two-way
synchronization is asking for trouble. This does mean all controls need
to be sent to the chromecast.